requirements.txt
google-generativeai>=0.4.0
pydantic>=2.0.0
pydantic-ai>=0.2.0
rich>=13.0.0
python-dotenv>=1.0.0
typer>=0.9.0
typing-extensions>=4.8.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
requests>=2.31.0
aiohttp>=3.9.0
